requirements.txt
Flask==3.0.0
line-bot-sdk==3.5.0
python-dotenv==1.0.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2 